17-12-25

+++ src/main/java/com/timesheetspro_api/common/model/CompanyEmployee/CompanyEmployee.java
@@ -178,7 +178,7 @@ public class CompanyEmployee implements UserDetails {
     private Integer lunchBreak;
 
     @Column(name = "working_hours_include_lunch", nullable = true)
-    private Integer workingHoursIncludeLunch;
+    private Float workingHoursIncludeLunch;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "ot_id ", referencedColumnName = "id")
